moved alternate links into a separate component so it'd be properly updated when navigating
